refactor waste fraction handler helpers
